+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ﻿# Changelog
 
+## [0.1.1] - 2025-5-18
+
+### Adding better priority management
+- StatePriority: allows different ways to manage where the state should be added.
+- PriorityManager: is responsible for managing all priorities, defining the order of the states.
+
 ## [0.0.1] - 2024-12-3
 
 ### This is the first release of *Unity Package master_state_machine*.